+++ butano/src/bn_affine_bg_attributes_hbe_handler.h
@@ -19,18 +19,18 @@ class affine_bg_attributes_hbe_handler
 {
 
 public:
-    static void setup_target(int target_id, iany& target_last_value)
+    static void setup_target(intptr_t target_id, iany& target_last_value)
     {
         target_last_value = last_value_type(target_id);
     }
 
-    [[nodiscard]] static bool target_visible(int target_id)
+    [[nodiscard]] static bool target_visible(intptr_t target_id)
     {
         auto handle = reinterpret_cast<void*>(target_id);
         return bgs_manager::hw_id(handle).has_value();
     }
 
-    [[nodiscard]] static bool target_updated(int target_id, iany& target_last_value)
+    [[nodiscard]] static bool target_updated(intptr_t target_id, iany& target_last_value)
     {
         last_value_type& last_value = target_last_value.value<last_value_type>();
         last_value_type new_value = last_value_type(target_id);
@@ -39,13 +39,13 @@ public:
         return updated;
     }
 
-    [[nodiscard]] static uint16_t* output_register(int target_id)
+    [[nodiscard]] static uint16_t* output_register(intptr_t target_id)
     {
         auto handle = reinterpret_cast<void*>(target_id);
         return hw::bgs::attributes_register(*bgs_manager::hw_id(handle));
     }
 
-    static void write_output_values(int target_id, const iany&, const void* input_values_ptr,
+    static void write_output_values(intptr_t target_id, const iany&, const void* input_values_ptr,
                                     uint16_t* output_values_ptr)
     {
         auto handle = reinterpret_cast<void*>(target_id);
@@ -53,11 +53,11 @@ public:
         bgs_manager::fill_hblank_effect_affine_attributes(handle, affine_bg_attributes_ptr, output_values_ptr);
     }
 
-    static void show(int)
+    static void show(intptr_t)
     {
     }
 
-    static void cleanup(int)
+    static void cleanup(intptr_t)
     {
         bgs_manager::reload();
     }
@@ -78,7 +78,7 @@ private:
         {
         }
 
-        explicit last_value_type(int target_id) :
+        explicit last_value_type(intptr_t target_id) :
             last_value_type(reinterpret_cast<void*>(target_id))
         {
         }
