+++ butano/src/bn_blending_transparency_attributes_hbe_handler.h
@@ -18,16 +18,16 @@ class blending_transparency_attributes_hbe_handler
 {
 
 public:
-    [[nodiscard]] static bool target_visible(int)
+    [[nodiscard]] static bool target_visible(intptr_t)
     {
         return true;
     }
 
-    static void setup_target(int, iany&)
+    static void setup_target(intptr_t, iany&)
     {
     }
 
-    [[nodiscard]] static bool target_updated(int, iany&)
+    [[nodiscard]] static bool target_updated(intptr_t, iany&)
     {
         BN_ASSERT(! display_manager::blending_fade_enabled(),
                    "Transparency and fade blendings can't be enabled at the same time");
@@ -35,22 +35,22 @@ public:
         return false;
     }
 
-    [[nodiscard]] static uint16_t* output_register(int)
+    [[nodiscard]] static uint16_t* output_register(intptr_t)
     {
         return hw::display::blending_transparency_register();
     }
 
-    static void write_output_values(int, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
+    static void write_output_values(intptr_t, const iany&, const void* input_values_ptr, uint16_t* output_values_ptr)
     {
         auto attributes_ptr = reinterpret_cast<const blending_transparency_attributes*>(input_values_ptr);
         display_manager::fill_blending_transparency_hblank_effect_attributes(attributes_ptr, output_values_ptr);
     }
 
-    static void show(int)
+    static void show(intptr_t)
     {
     }
 
-    static void cleanup(int)
+    static void cleanup(intptr_t)
     {
         display_manager::reload_blending_transparency();
     }
